requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
opencv-python==4.10.0.84
numpy==2.0.2
Pillow==10.4.0
